requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
sqlalchemy==2.0.7
pydantic==1.10.7
python-dotenv==1.0.0
alpaca-trade-api==3.0.0
pandas==2.0.0
numpy==1.24.2
scikit-learn==1.2.2
matplotlib==3.7.1
python-jose==3.3.0
passlib==1.7.4
pytest==7.3.1
httpx==0.24.0
requests==2.28.2
bcrypt==4.0.1
asyncio==3.4.3
websockets<11.0.0,>=9.0.0
aiohttp==3.8.1
psycopg2-binary==2.9.6
alembic==1.10.3
textblob==0.17.1
nltk==3.8.1
tensorflow==2.12.0
keras==2.12.0
statsmodels==0.14.0
yfinance==0.2.18
ta==0.10.2
